pyproject.toml
[project]
name = "hsn-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "pandas>=2.2.3",
]
